# src/main/resources/assets/api-docs.yaml
swagger: '2.0'
info:
  title: Consent
  description: A set of web-services containing methods for ingesting and retrieving expressions representing data use restrictions found within consent forms.
  version: 1.0.1
  license:
    name: BSD
    url: 'http://opensource.org/licenses/BSD-3-Clause'
  termsOfService: 'https://github.com/DataBiosphere/consent'
securityDefinitions:
  consent_auth:
    type: oauth2
    authorizationUrl: 'https://accounts.google.com/o/oauth2/auth'
    flow: implicit
    scopes:
      profile: profile
      email: email
paths:
...
  /api/consent/associations/{associationType}/{id}:
    get:
      summary: getConsentsForAssociation
      description: Returns a List with the Consent IDs that are related to the specified association
      parameters:
        - name: associationType
          in: path
          description: The association type
          required: true
          type: string
        - name: id
          in: path
          description: The id of the association
          required: true
          type: string
      tags:
        - Consent
        - Association
      responses:
        200:
          description: A list of Consent IDs
          schema:
            type: array
            items:
              type: string
        400:
          description: Required parameter is null
        default:
          description: Unexpected error
  /api/consent/{consentId}/association:
    get:
      summary: Get Associations for Consent
      description: Returns the associations identified by the parameters.
      parameters:
        - name: consentId
          in: path
          description: Consent ID
          required: true
          type: string
        - name: associationType
          in: query
          description: "Association's type, 'sample', 'sampleSet', or 'workspace'"
          required: true
          type: string
        - name: id
          in: query
          description: Optional ID of the requested association
          required: false
          type: string
      tags:
        - Consent
        - Association
      responses:
        200:
          description: An array of Consent Associations
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
        400:
          description: Required parameter is null
        default:
          description: Unexpected error
    post:
      summary: Create Associations for Consent
      description: Create the associations sent as parameters.
      parameters:
        - name: consentId
          in: path
          description: Consent ID
          required: true
          type: string
        - name: body
          in: body
          description: Associations to be created
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
      tags:
        - Consent
        - Association
      responses:
        200:
          description: An array of created consent associations
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
        default:
          description: Unexpected error
    put:
      summary: Update Associations for Consent
      description: Update the association identified by the parameters
      parameters:
        - name: consentId
          in: path
          description: Consent ID
          required: true
          type: string
        - name: body
          in: body
          description: Associations to be updated
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
      tags:
        - Consent
        - Association
      responses:
        200:
          description: An array with the updated consent associations
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
        default:
          description: Unexpected error
    delete:
      summary: Delete Associations for Consent
      description: Deletes the association identified by the parameters
      parameters:
        - name: consentId
          in: path
          description: Consent ID
          required: true
          type: string
        - name: associationType
          in: query
          description: "Association's type"
          required: true
          type: string
        - name: id
          in: query
          description: Object ID of the requested association
          required: true
          type: string
      tags:
        - Consent
        - Association
      responses:
        200:
          description: An array with the deleted consent associations
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsentAssociation'
        default:
          description: Unexpected error
...
  /api/consent:
    get:
      summary: Find Consent by name
      description: Returns the Consent identified by the name.
      produces:
        - application/json
      parameters:
        - name: name
          in: query
          description: Consent Name
          required: true
          type: string
      tags:
        - Consent
      responses:
        200:
          description: Returns the requested consent.
          schema:
            $ref: '#/definitions/Consent'
        400:
          description: The consent associated with the provided name has not been approved.
        404:
          description: The consent associated with the provided name could not be found.
    post:
      summary: Create new Consent
      description: Creates the consent sent with the request.
      parameters:
        - name: rec
          in: body
          description: Object that represents a Consent
          required: true
          schema:
            $ref: '#/definitions/Consent'
      tags:
        - Consent
      responses:
        201:
          description: Returns the URL to find the created consent.
        500:
          description: Internal Server Error.
  /api/consent/{id}:
    get:
      summary: Get Consent by ID
      description: Returns the consent identified by the ID.
      parameters:
        - name: id
          in: path
          description: Consent ID
          required: true
          type: string
      tags:
        - Consent
      responses:
        200:
          description: Returns the requested consent.
          schema:
            $ref: '#/definitions/Consent'
        404:
          description: The consent associated with the ID couldn't be found.
    put:
      summary: Update Consent by ID
      description: Updates the consent identified by the ID.
      parameters:
        - name: id
          in: path
          description: Consent ID
          required: true
          type: string
        - name: updated
          in: body
          description: Object that represents the consent to be updated
          required: true
          schema:
            $ref: '#/definitions/Consent'
      tags:
        - Consent
      responses:
        200:
          description: Returns the requested consent.
          schema:
            $ref: '#/definitions/Consent'
        400:
          description:
            Consent updates are only allowed when
              1-There have been no previous elections created for this consent
              2-The most recent election has been closed/cancelled AND archived
        404:
          description: The consent associated with the ID couldn't be found.
        500:
          description: Internal server error
...
  /api/consents:
    get:
      summary: findByIds
      description: Returns the consents required in the parameter list
      parameters:
        - name: ids
          in: query
          description: List of consent IDs
          required: true
          type: array
          items:
            type: string
      tags:
        - Consent
      responses:
        200:
          description: Returns the list of found consents.
          schema:
            type: array
            items:
              $ref: '#/definitions/Consent'
        404:
          description: The consent IDs list wasn't send, or some consents couldn't be found.
  /api/consents/{associationType}:
    get:
      summary: findByAssociationType
      description: Returns the consents that have associated a record in consentassociation that matches the requested association type.
      parameters:
        - name: associationType
          in: path
          description: Association Type.
          required: true
          type: string
      tags:
        - Consent
      responses:
        200:
          description: Returns the list of found consents.
          schema:
            type: array
            items:
              $ref: '#/definitions/Consent'
        404:
          description: The association type wasn't send, or the application couldn't find any consents that matched the search criteria.
...
  /api/dar/find/{id}/consent:
    get:
      summary: describeConsentForDAR
      description: Returns the consent related to this DAR.
      parameters:
        - name: id
          in: path
          description: The id of the DAR
          required: true
          type: string
      tags:
        - Data Access Request
        - Consent
      responses:
        200:
          description: Returns the Consent.
          schema:
            $ref: '#/definitions/Consent'
        404:
          description: The requested couldn't be found or the DAR doesn't exist.
...
definitions:
...
  ConsentAssociation:
    type: object
    properties:
      associationType:
        type: string
        description: 'Defines the type of the association. Possible values: sample, sampleSet, workspace.'
      elements:
        type: array
        items:
          type: string
        description: Elements, in a hash set, to avoid duplicates.
  Consent:
    type: object
    properties:
      consentId:
        type: string
        description: The ID of the entity.
      requiresManualReview:
        type: boolean
        description: Describes if this consent has to be reviewed manually.
      dataUseLetter:
        type: string
        description: The URL of the saved file.
      useRestriction:
        type: string
        description: The UseRestriction of this consent. Used to match against another entities.
      dataUse:
        type: string
        description: The Data Use question/answer set for this consent.
      name:
        type: string
        description: Name to identify the consent.
      dulName:
        type: string
        description: Name that identifies the Data Use Letter file.
      createDate:
        type: string
        format: date
        description: Creation Date.
      lastUpdate:
        type: string
        format: date
        description: Date of the last update.
      sortDate:
        type: string
        format: date
        description: The lastest modification date.
      translatedUseRestriction:
        type: string
        format: date
        description: The translated use restriction of the consent.
      lastElectionStatus:
        type: string
        description: Displays the associated election if this exists, else is null.
      lastElectionArchived:
        type: boolean
        description: Flag indicating if the last election associated is archived or not. Null if election doesn't exists
      groupName:
        type: string
        description: The Group Name provided by ORSP
      updateStatus:
        type: boolean
        description: Indicates if a consent has been updated from orsp
...